pyproject.toml
[project]
name = "local-llm"
version = "0.1.0"
description = "Local RAG assistant for FIR dissertations using LangChain, Qdrant, and Ollama."
requires-python = ">=3.10"
dependencies = [
  "typer[all]>=0.12",
  "pydantic-settings>=2.7",
  "langchain>=0.3.9",
  "langchain-community>=0.3.9",
  "langchain-openai>=0.2.0",
  "langchain-qdrant>=0.2.0",
  "qdrant-client>=1.12.1",
  "sentence-transformers>=3.0.0",
  "pdfplumber>=0.11.0",
  "streamlit>=1.34.0",
  "fastapi>=0.115.0",
  "uvicorn[standard]>=0.30.0",
  "httpx>=0.27.0",
  "rich>=13.0.0",
]

[project.optional-dependencies]
dev = [
  "pytest>=8.0.0",
  "ruff>=0.4.0",
]

[project.scripts]
local-llm = "localllm.cli:app"

[build-system]
requires = ["setuptools>=68.0"]
build-backend = "setuptools.build_meta"

[tool.setuptools.packages.find]
where = ["src"]

[tool.ruff]
line-length = 100
target-version = "py310"

